requirements.txt
# AWS Bedrock and AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Vector Store and Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0  # Use numpy 1.x for compatibility
jsonlines>=4.0.0

# RAG and Evaluation
ragas>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-aws>=0.1.0  # For Bedrock (answer generation)
langchain-google-genai>=2.0.0  # For Gemini (Ragas evaluation only)

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
pyyaml>=6.0.1

# Development
jupyter>=1.0.0
ipykernel>=6.29.0
black>=24.0.0
pytest>=7.4.0

